controller for downloading kda information created and js script writed to show win and kda

+++ src/main/java/pl/noip/lolstats/lol/stats/dto/Match.java
@@ -22,5 +22,5 @@ public class Match {
     private String gameDuration;
     private Boolean win;
     private String participantId;
-    private int kda;
+    private double kda;
 }
